requirements.txt
python-dotenv
langchain-community
langchain-groq
pandas
streamlit
mysql-connector-python

